NotificationBar: Force DateView to collapse to gain space for clear_all_button
When DateFormat.getLongDateFormat() text is too long, the notification "clear_all_button" button
is hidden because DateView width. This change redesign the notification bar layout to force
DateView to collapse. This forces you to change the position of the buttons to the right.

Alternatives:

Patch 1: Align buttons (rotation and settings) to the right

Patch 2: Align buttons (rotation and settings) to the middle

Patch 3: Compute the text width of DateView to fit the maximum allowed by the status bar.
           * Calculate dinamicaly the width of the DateView to assign the maximum width
             when it fit in the screen, and stretch when not fit
           * Redesign the DateView widget. Now is a LinearLayout with 2 TextView's. This allow
             the use of ellipsize, because it is set to single line. Otherwise when the text is set
             the DateView don't show the ellipsize, although the view fit to the width.
Patch 4: Fix bug that in some circumstances ellipsize is not displayed

Change-Id: I2486c5e6b7d83a3c483519c17cdf57e90b50c79e

Conflicts:

	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java